+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Miscellaneous
+- Add tests
+    - positive/negative tests for action
+
 ## [1.0.1] - 2018-02-10
 
 ### Miscellaneous